requirements.txt
flask==2.2.3
gunicorn==20.1.0
python-telegram-bot==13.5
Werkzeug~=2.2.3
pymongo~=4.3.3